requirements.txt
streamlit>=1.28.0
azure-cognitiveservices-speech>=1.31.0
requests>=2.31.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-cov>=4.1.0